+++ index.html
@@ -236,7 +236,6 @@
     </template>
 
   <script src="./js/main.js" type="module"></script>
-  <script src="./js/functions.js"></script>
 
 </body>
 </html>
